plantation calculate

+++ src/main/java/tr/ozanbey/agricalc/webapp/service/domain/City.java
@@ -39,7 +39,7 @@ public class City extends AbstractEntity {
     private BigDecimal fuelPrice;
 
     @Column(name = "electricity")
-    private Double electricity;
+    private BigDecimal electricity;
 
     public City(Long selectedCityId) {
         super.setId(selectedCityId);
